% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carveFastQL.R
\name{carveFastQL}
\alias{carveFastQL}
\title{carveFastQL
process a text file containing a fastql script
returns a set of wait times}
\usage{
carveFastQL(loc)
}
\arguments{
\item{loc}{the location of the script file.}
}
\description{
carveFastQL
process a text file containing a fastql script
returns a set of wait times
}
