% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collator.R
\name{collator}
\alias{collator}
\title{collator
compiles gas exchange, spectroscopic and fluorescence data into a set of graphs and 
tables for use in rmarkdown reports}
\usage{
collator(ge, ecs, fop)
}
\arguments{
\item{ge}{the gas exchange data to use}

\item{ecs}{the spec data to use}

\item{fop}{the fluorescence data to use}
}
\description{
collator
compiles gas exchange, spectroscopic and fluorescence data into a set of graphs and 
tables for use in rmarkdown reports
}
