% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutter.R
\name{cutter}
\alias{cutter}
\title{Cutter
Divide a dataframe into parts based on two ratios.}
\usage{
cutter(df, r1, r2)
}
\arguments{
\item{df}{the dataframe to be divided}

\item{r1}{the first chunk size}

\item{r2}{the second chunk size}
}
\description{
Cutter
Divide a dataframe into parts based on two ratios.
}
