% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table_old}
\alias{format_table_old}
\title{format_table_old
im not sure this will work if it's not in the global env actually...}
\usage{
format_table_old(num)
}
\arguments{
\item{num}{index to use.}
}
\description{
format_table_old
im not sure this will work if it's not in the global env actually...
}
