% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitter}
\alias{splitter}
\title{splitter
divide a df into a list, with each element containing all records in between two paired times.}
\usage{
splitter(df, timer, timer2)
}
\arguments{
\item{df}{the dataframe to divide}

\item{timer}{the first vector of times to select by}

\item{timer2}{the second vector of times to select by}
}
\description{
splitter
divide a df into a list, with each element containing all records in between two paired times.
}
