% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemelt.R
\name{timemelt}
\alias{timemelt}
\title{timemelt
just a convenient wrapper around reshape2::melt}
\usage{
timemelt(df)
}
\arguments{
\item{df}{the dataframe to melt.}
}
\description{
timemelt
just a convenient wrapper around reshape2::melt
}
